pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "Add description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "matplotlib>=3.10.7",
    "neo4j>=6.0.3",
    "networkx>=3.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
